Melee enemy complete.
Melee enemy is now fully functioning and animated in all four directions. All elements of the enemy can be manually adjusted - note, do NOT change "Time To Attack" or "Attack Duration" without also changing the animation clips. They are specifically timed to give the player a chance to learn when an attack will happen.