Collectibles and UI messages created
Note: With this patch, the player NO LONGER HAS DASH OR PARRY ENABLED BY DEFAULT AND REQUIRES THE PICKUPS TO USE THEM. In addition, do NOT attempt to transfer these new elements to the main scene yet, as they are currently incomplete.

Collectibles of all 3 kinds created - Dash pickup, Parry pickup, and standard pickups. At the moment, they are NOT prefabs, and exist only in the Greybox scene.

This is in part because they directly reference a new UI component, which showcases messages to the player - including the base instructions for their abilities upon receiving them.